New UI Page & Button Functionality
Added a new page which gets launched on the "+" button click and takes input from user and creates a  new group and displays the newly created group name in the group page.  However this newly created group doesn't reload on a new  launch since we do not have any database associated with the project and our app loads a static list as of now.

Also added the functionality to the buttons in the header based the page the app is currently in.